+++ examples/Markov/Project.toml
@@ -2,3 +2,7 @@ name = "Markov"
 uuid = "dac345a6-b16d-456a-a2c5-d9fc9368bae6"
 authors = ["Evan Patterson <evan@epatters.org>"]
 version = "0.1.0"
+
+[deps]
+Catlab = "134e5e36-593f-5add-ad60-77f754baafbe"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
