+++ src/Catlab.jl
@@ -6,6 +6,5 @@ include("wiring_diagrams/WiringDiagrams.jl")
 include("graphics/Graphics.jl")
 include("programs/Programs.jl")
 include("linear_algebra/LinearAlgebra.jl")
-include("experimental/Experimental.jl")
 
 end
